admin: highlight sidebar sources/destinations when browsing connectors
This commit highlights the "Sources"/"Destinations" entries in the sidebar based
on the role when the user is browsing the list of connectors or adding a new
connection.
